requirements.txt
Flask
fastapi
uvicorn
neo4j
sentence-transformers
langchain
langchain-google-genai
langchain-experimental
langchain-core
langchain-neo4j
python-dotenv
requests
pydantic
pypdf2
typing


